Updated interact header function signature to match source

+++ include/utils/terminal.h
@@ -10,6 +10,6 @@
 void typewrite(char *text, int delay, ...);
 void get_input(char *prompt, char *input, size_t buffer_size, uint16_t delay, bool to_lower);
 void clear_terminal(void);
-void interact(char *prompt, int delay);
+void interact(char *prompt, uint16_t delay);
 
 #endif
